refactor: refactor dragNDrop
