+++ apps/cowswap-frontend/src/modules/hooksStore/containers/PreHookButton/index.tsx
@@ -1,7 +1,6 @@
 import PLUS_ICON from '@cowprotocol/assets/cow-swap/plus.svg'
 import { useWalletInfo } from '@cowprotocol/wallet'
 
-import { useDrop } from 'react-dnd'
 import SVG from 'react-inlinesvg'
 
 import * as styledEl from './styled'
@@ -11,10 +10,6 @@ import { useRemoveHook } from '../../hooks/useRemoveHook'
 import { useReorderHooks } from '../../hooks/useReorderHooks'
 import { AppliedHookItem } from '../../pure/AppliedHookItem'
 import { HookTooltip } from '../../pure/HookTooltip'
-
-const ItemTypes = {
-  HOOK: 'hook',
-}
 
 export interface PreHookButtonProps {
   onOpen(): void
@@ -27,10 +22,8 @@ export function PreHookButton({ onOpen, onEditHook }: PreHookButtonProps) {
   const removeHook = useRemoveHook()
   const moveHook = useReorderHooks('preHooks')
 
-  const [, drop] = useDrop(() => ({ accept: ItemTypes.HOOK }))
-
   return (
-    <div ref={drop}>
+    <>
       {preHooks.length > 0 && (
         <styledEl.HookList>
           {preHooks.map((hookDetails, index) => (
@@ -53,6 +46,6 @@ export function PreHookButton({ onOpen, onEditHook }: PreHookButtonProps) {
           <SVG src={PLUS_ICON} /> Add Pre-Hook Action <HookTooltip isPreHook />
         </styledEl.AddHookButton>{' '}
       </styledEl.Wrapper>
-    </div>
+    </>
   )
 }
